Add global variables for Screen Size

+++ BreakOut-Sigil/src/GameManagement/Utilities.h
@@ -3,8 +3,8 @@
 #include "GameManagement/Color.h"
 struct Vector2
 {
-	int x{};
-	int y{};
+	float x{};
+	float y{};
 };
 
 struct Rect
